refactor: rename directory components to functions
