+++ js/functions/check-symbol.js
