+++ js/functions/set-settings.js
