+++ js/functions/switch-lang.js
